requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
Pillow==6.0.0
psycopg2-binary==2.8.2
pycparser==2.19
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
